backend/src/requirements.txt
openai
python-dotenv
fastapi 
uvicorn 
python-multipart
pytest